requirements.txt
# Core
httpx>=0.25.0
requests>=2.31.0

# Twitter & Reddit
tweepy>=4.14.0
praw>=7.0.0

# RSS & Parsing
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# AI/LLM
openai>=1.3.0

# Config & Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0
pytz>=2023.3
loguru>=0.7.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
